% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{add_noise_to_matrix}
\alias{add_noise_to_matrix}
\title{Add noise or create noise matrix}
\usage{
add_noise_to_matrix(x, learning_rate = 0.9, merge = TRUE)
}
\arguments{
\item{x}{matrix}

\item{learning_rate}{numeric from 0 to 1, determining probability that each element in the row is sampled perfectly}

\item{merge}{logical default is TRUE and return original matrix multiplied by noise matrix. FALSE will return only the noise matrix}
}
\value{

}
\description{
Add noise or create noise matrix
}
\examples{

}
