% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{make_AB_timeline}
\alias{make_AB_timeline}
\title{Make AB timeline}
\usage{
make_AB_timeline(
  num_trials = 5,
  events = c("A", "B"),
  A_onsets = 1,
  A_durations = 50,
  B_onsets = 50,
  B_durations = 50
)
}
\arguments{
\item{num_trials}{integer, the number of trials to add to timeline}

\item{events}{character vector, defaults to A and B events}

\item{A_onsets}{integer vector, the onset times for each A event on each trial}

\item{A_durations}{integer vector, the duration times for each A event on each trial}

\item{B_onsets}{integer vector, the onset times for each B event on each trial}

\item{B_durations}{integer vector, the duration times for each B event on each trial}
}
\value{
data.frame
}
\description{
Make AB timeline
}
\examples{
make_AB_timeline(
 num_trials = 5,
 events = c('A', 'B'),
 A_onsets = 1,
 A_durations = 50,
 B_onsets = round(rnorm(5,100,10)),
 B_durations = 50
)
}
