% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{make_event_vectors}
\alias{make_event_vectors}
\title{Make Event Vectors}
\usage{
make_event_vectors(event_names = LETTERS[1:10], vector_length = 10)
}
\arguments{
\item{event_names}{character vector, names for each event}

\item{vector_length}{integer, number of 1s for each vector}
}
\value{
matrix
}
\description{
Make Event Vectors
}
\examples{
make_event_vectors()
}
