% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{make_temporal_vectors}
\alias{make_temporal_vectors}
\title{Make Temporal Vectors}
\usage{
make_temporal_vectors(num_timesteps = 200, overlap = 2)
}
\arguments{
\item{num_timesteps}{integer, the number of timesteps within a trial}

\item{overlap}{integer, controls overlap between vectors}
}
\value{
matrix, each row is a temporal vector
}
\description{
Make Temporal Vectors
}
\examples{
make_temporal_vectors(10,2)
}
