% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{make_trial}
\alias{make_trial}
\title{Make a trial}
\usage{
make_trial(trial_num, event_names, onsets, durations)
}
\arguments{
\item{trial_num}{integer, the trial number}

\item{event_names}{character vector, names from event_vectors matrix for the events that appear in the trial}

\item{onsets}{integer vector, onset times for each event listed in event_names}

\item{durations}{integer vector, durations for each event listed in event_names}
}
\value{
data.frame
}
\description{
Make a trial
}
\examples{
make_trial(1,event_names=c("A","B"),c(1,100),duration=50)
}
