% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{timeline_to_vector}
\alias{timeline_to_vector}
\title{Convert timeline to lists of environment vectors}
\usage{
timeline_to_vector(timeline, event_vectors, temporal_vectors, timesteps = 200)
}
\arguments{
\item{timeline}{dataframe, created by make_timeline()}

\item{event_vectors}{matrix, created by make_event_vectors()}

\item{temporal_vectors}{matrix, created by make_temporal_vectors()}

\item{timesteps}{integer, number of timesteps per trial}
}
\value{
list, containing a environment vector matrix for each trial.
}
\description{
Convert timeline to lists of environment vectors
}
