% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{timeline_to_vector_noise}
\alias{timeline_to_vector_noise}
\title{Convert timeline to environment vectors with noise}
\usage{
timeline_to_vector_noise(
  timeline,
  event_vectors,
  temporal_vectors,
  timesteps = 200,
  noise = TRUE,
  L = 0.1
)
}
\arguments{
\item{timeline}{dataframe, created by make_timeline()}

\item{event_vectors}{matrix, created by make_event_vectors()}

\item{temporal_vectors}{matrix, created by make_temporal_vectors()}

\item{timesteps}{integer, number of timesteps per trial}
}
\value{
list, containing a environment vector matrix for each trial.
}
\description{
Convert timeline to environment vectors with noise
}
